Fix CR3BP benchmark assertion error by denormalizing halo energies
- Fixed tensor dimension mismatch in diffusion model sampling
- Added denormalization of halo energies from [0,1] to physical range [0.008, 0.095]
- Updated benchmark to handle normalized vs physical halo energy values
- Added proper error handling for CR3BP simulation failures
- Benchmark now runs successfully without assertion errors

+++ Reflected-Diffusion/sde_lib.py
@@ -136,7 +136,7 @@ class RVESDE(SDE):
         sigma = self.sigma_min * (self.sigma_max / self.sigma_min) ** t
         drift = torch.zeros_like(x)
         diffusion = sigma * torch.sqrt(torch.tensor(2 * (np.log(self.sigma_max) - np.log(self.sigma_min)),
-                                                    device=t.device))
+                                                    device=t.device, dtype=torch.float32))
         return drift, diffusion
 
     def marginal_prob(self, x, t):
